feature: ledger integration (#955)
* Add connection ledger UI and update signing flow for using the ledger

* Add animation for ledger, refactor code for dark mode

* Remove unused imports in avatar and ledger accounts list

* Update cursor behavior and colors for disabled checkboxes

The code modifies the Cursor behavior and colors for disabled checkboxes and buttons. It also introduces a check to prevent duplicate Ledger accounts from being connected by deactivating the selection and changing the cursor to 'not-allowed' when an account is already connected. The checkbox colors have been adjusted accordingly.

* Adjust spinner animation colors and tidy up layout styling

* Update how imported account visibility is checked in tests

Replaced the method of checking the visibility of imported accounts in e2e tests. Instead of searching for the exact text 'Imported', the test now looks for a specific dataTestId ('import-account-icon'). Also, added the 'import-account-icon' dataTestId to the SvgIcon component in the hash.tsx file.

* Update SVG illustration for ledger

* Add Ledger integration business logic

* Update Ledger integration UI

* Refactor account import functionality from ledger (#984)

* Update maxHeight property in account-list component (#985)

* Add useEffect to scrollToTop in Ledger error components (#986)

* Remove lodash.debounce (#987)

* Handle issue with browser's device permission window

* Update global styles and layout parameters

This commit includes a change in the global styles file to allow elements to display in a flex layout and centrally align their content.
In addition, modifications have been made to the layout-window file to ensure elements take up 100% width, enhancing the app's layout flexibility.

* Fix small issues

---------

Co-authored-by: Dmytro Vynnyk <simbiatoff@gmail.com>