feature: improve extension in-tab view (#988)
* Improve extension in-tab view

* Update Overlay component and add delays in e2e tests

The Overlay component's position and size properties were adjusted to fit its new design requirements. The width is now fixed, and it is centered using transform. Meanwhile, explicit delays were added in the end-to-end tests for 'buy-cspr.spec.ts' to ensure proper closing of modal windows before continuing with the tests.